configured bable and some minor changes in the index.js
