+++ package.json
@@ -5,20 +5,16 @@
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "watch": "jest --watch *.js",
+    "start": "webpack-dev-server --mode development --entry ./src/index.js --output-filename ./dist/main.js",
     "build": "webpack",
-    "babel": {
-      "env": {
-        "test": {
-          "presets": [
-            "@babel/preset-env"
-          ]
-        }
-      }
-    }
+    "watch": "jest --watch *.js"
   },
   "jest": {
-    "verbose": true
+    "verbose": true,
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less|scss|sass)$": "identity-obj-proxy"
+    }
   },
   "repository": {
     "type": "git",
@@ -32,16 +28,22 @@
   },
   "homepage": "https://github.com/Abdusaid10/testing-tic-tac-toe#readme",
   "dependencies": {
+    "babel-loader": "^8.1.0",
     "lodash": "^4.17.15",
     "webpack": "^4.42.1"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
     "@babel/preset-env": "^7.9.0",
+    "babel-jest": "^25.2.6",
     "babel-preset-env": "^1.7.0",
+    "css-loader": "^3.5.1",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^25.2.7",
+    "style-loader": "^1.1.3",
     "webpack-cli": "^3.3.11"
   }
 }
